[Ruchir]Added: Verify user and check user api

+++ src/main/java/com/bridgelabz/userregistrationproblem/repository/UserRegistrationRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bridgelabz.userregistrationproblem.entity.UserEntity;
 
-public interface UserRegistrationRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRegistrationRepository extends JpaRepository<UserEntity, Long> {
 
 	Optional<UserEntity> findByEmailId(String emailId);
 }
